Split dataset loading into separate assets
