--- src/mozilla_sec_eia/models/sec10k/ex_21/data.py
@@ -1,1 +1,0 @@
-"""Define methods and assets for handling datasets used by."""
