+++ tests/unit/models/sec10k/ex21_model_test.py
@@ -11,12 +11,12 @@ from mozilla_sec_eia.library.validation_helpers import (
     pandas_compute_precision_recall,
     strip_down_company_names,
 )
-from mozilla_sec_eia.models.sec10k.ex_21.ex21_validation_helpers import (
-    clean_ex21_validation_set,
-)
-from mozilla_sec_eia.models.sec10k.ex_21.inference import (
+from mozilla_sec_eia.models.sec10k.ex_21.data.common import (
     LABELS,
     get_flattened_mode_predictions,
+)
+from mozilla_sec_eia.models.sec10k.ex_21.ex21_validation_helpers import (
+    clean_ex21_validation_set,
 )
 from mozilla_sec_eia.models.sec10k.utils.layoutlm import get_id_label_conversions
